requirements.txt
pytest>=7.0.0
selenium>=4.0.0
allure-pytest
allure-python-commons
